backend-archive/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "promotor"
version = "0.1.0"
description = "Bloomberg Terminal-style Beauty Brand Promotion Manager powered by AI Agents"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [
    { name = "Promotor Team" }
]
dependencies = [
    # LangChain & LangGraph
    "langchain>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "langsmith>=0.1.0",

    # FastAPI & Web
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "websockets>=11.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",

    # Database & Cache
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.27.0",
    "redis>=4.5.0",
    "supabase>=2.0.0",

    # Task Queue
    "celery[redis]>=5.3.0",

    # Web Scraping (Stealth)
    "playwright>=1.40.0",
    "playwright-stealth>=1.0.0",
    "httpx>=0.24.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.3.0",

    # ML & NLP
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "sentence-transformers>=2.2.0",
    "prophet>=1.1.0",
    "statsmodels>=0.14.0",
    "scikit-learn>=1.3.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",

    # Vector Store
    "pinecone-client>=2.2.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "structlog>=23.1.0",
    "tenacity>=8.2.0",
    "croniter>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.4.0",
]

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
